chore(tests): add tests from header builder util
