+++ src/utils/mod.rs
@@ -1,39 +1,6 @@
-use axum::http::{HeaderMap, HeaderValue};
-use reqwest::{StatusCode, header};
-
-use crate::utils::rand_ua::randomize_user_agent;
+pub mod build_headers;
 pub mod rand_ua;
-
-/// Panics if:
-/// - The `randomize_user_agent` function panics due to an empty list of User-Agent strings or a failure in random selection.
-/// - The `HeaderValue::from_str` method panics if the provided User-Agent string is invalid.
-/// - The `HeaderValue::from_static` method panics if the provided static string is invalid (unlikely with hardcoded values).
-pub fn build_headers(endpoint: &str) -> HeaderMap {
-    let user_agent = randomize_user_agent(endpoint);
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        header::USER_AGENT,
-        HeaderValue::from_str(&user_agent).expect("Invalid user agent"),
-    );
-    headers.insert(
-        header::ACCEPT,
-        HeaderValue::from_static(
-            "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
-        ),
-    );
-    headers.insert(
-        header::ACCEPT_LANGUAGE,
-        HeaderValue::from_static("en-US,en;q=0.5"),
-    );
-    headers.insert("Cache-Control", HeaderValue::from_static("no-cache"));
-
-    headers.insert(
-        header::REFERER,
-        HeaderValue::from_str("https://www.google.com/").unwrap(),
-    );
-
-    headers
-}
+use reqwest::StatusCode;
 
 /// Handles all `reqwest` errors by categorizing them based on their kind.
 ///
